pyproject.toml
[project]
name = "swarm-intelligence-simulation"
version = "0.1.0"
description = "Multi-Agent Search & Rescue with MADDPG and GNN Communication"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.0",
    "pettingzoo>=1.25.0",
    "pillow>=12.1.0",
    "pygame>=2.6.1",
    "scipy>=1.16.3",
    "torch>=2.9.1",
    "matplotlib>=3.8.0",
    "tensorboard>=2.16.0",
    "pyyaml>=6.0.1",
    "tqdm>=4.66.0",
]
